+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] - 2020-02-15
+### Changed
+- A Factory class is now being used to instantiate all of the IP rules, instead
+  of include files.
+
 ## [0.2.1] - 2020-02-14
 ### Changed
 - All of the IANA conversion logic has been moved to the [IP Analysis Helper] project.
@@ -29,8 +34,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Basic IP Analysis tools
 - CSV parser for IANA IPv[46] Special-Purpose Address Registry
 
-[Unreleased]: https://github.com/dharple/ip-analysis/compare/v0.2.1...develop
-[0.2.0]: https://github.com/dharple/ip-analysis/compare/v0.2.0...0.2.1
+[Unreleased]: https://github.com/dharple/ip-analysis/compare/v0.3.0...develop
+[0.3.0]: https://github.com/dharple/ip-analysis/compare/v0.2.1...0.3.0
+[0.2.1]: https://github.com/dharple/ip-analysis/compare/v0.2.0...0.2.1
 [0.2.0]: https://github.com/dharple/ip-analysis/compare/v0.1.0...0.2.0
 [0.1.0]: https://github.com/dharple/ip-analysis/releases/tag/v0.1.0